# docs/thv-registry-api/swagger.yaml
components:
  schemas:
...
    model.Icon:
      properties:
        mimeType:
          example: image/png
          type: string
        sizes:
          items:
            type: string
          type: array
          uniqueItems: false
        src:
          example: https://example.com/icon.png
          format: uri
          maxLength: 255
          type: string
        theme:
          type: string
      type: object
...
    model.Status:
      type: string
      x-enum-varnames:
      - StatusActive
      - StatusDeprecated
      - StatusDeleted
    model.Transport:
      description: Transport is required and specifies the transport protocol configuration
      properties:
        headers:
          items:
            $ref: '#/components/schemas/model.KeyValueInput'
          type: array
          uniqueItems: false
        type:
          example: stdio
          type: string
        url:
          example: https://api.example.com/mcp
          type: string
      type: object
    v0.Metadata:
      properties:
        count:
          type: integer
        nextCursor:
          type: string
      type: object
    v0.RegistryExtensions:
      properties:
        isLatest:
          type: boolean
        publishedAt:
          format: date-time
          type: string
        status:
          $ref: '#/components/schemas/model.Status'
        updatedAt:
          format: date-time
          type: string
      type: object
    v0.ResponseMeta:
      properties:
        io.modelcontextprotocol.registry/official:
          $ref: '#/components/schemas/v0.RegistryExtensions'
      type: object
    v0.ServerJSON:
      properties:
        $schema:
          example: https://static.modelcontextprotocol.io/schemas/2025-10-17/server.schema.json
          format: uri
          minLength: 1
          type: string
        _meta:
          $ref: '#/components/schemas/v0.ServerMeta'
        description:
          example: MCP server providing weather data and forecasts via OpenWeatherMap
            API
          maxLength: 100
          minLength: 1
          type: string
        icons:
          items:
            $ref: '#/components/schemas/model.Icon'
          type: array
          uniqueItems: false
        name:
          example: io.github.user/weather
          maxLength: 200
          minLength: 3
          pattern: ^[a-zA-Z0-9.-]+/[a-zA-Z0-9._-]+$
          type: string
        packages:
          items:
            $ref: '#/components/schemas/model.Package'
          type: array
          uniqueItems: false
        remotes:
          items:
            $ref: '#/components/schemas/model.Transport'
          type: array
          uniqueItems: false
        repository:
          $ref: '#/components/schemas/model.Repository'
        title:
          example: Weather API
          maxLength: 100
          minLength: 1
          type: string
        version:
          example: 1.0.2
          type: string
        websiteUrl:
          example: https://modelcontextprotocol.io/examples
          format: uri
          type: string
      type: object
    v0.ServerListResponse:
      properties:
        metadata:
          $ref: '#/components/schemas/v0.Metadata'
        servers:
          items:
            $ref: '#/components/schemas/v0.ServerResponse'
          type: array
          uniqueItems: false
      type: object
    v0.ServerMeta:
      properties:
        io.modelcontextprotocol.registry/publisher-provided:
          additionalProperties: {}
          type: object
      type: object
    v0.ServerResponse:
      properties:
        _meta:
          $ref: '#/components/schemas/v0.ResponseMeta'
        server:
          $ref: '#/components/schemas/v0.ServerJSON'
      type: object
  securitySchemes:
    BearerAuth:
      description: 'OAuth 2.0 Bearer token authentication. Format: "Bearer {token}"'
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    url: https://github.com/stacklok/toolhive
  description: |-
    API for accessing MCP server registry data and deployed server information
    This API provides endpoints to query the MCP (Model Context Protocol) server registry,
    get information about available servers, and check the status of deployed servers.

    Authentication is required by default. Use Bearer token authentication with a valid
    OAuth/OIDC access token. The /.well-known/oauth-protected-resource endpoint provides
    OAuth discovery metadata (RFC 9728).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: ToolHive Registry API
  version: "0.1"
openapi: 3.1.0
paths:
  /{registryName}/v0.1/servers/{serverName}/versions/{version}:
    delete:
      description: Delete a server version from a specific managed registry
      parameters:
      - description: Registry name
        in: path
        name: registryName
        required: true
        schema:
          type: string
      - description: Server name (URL-encoded)
        in: path
        name: serverName
        required: true
        schema:
          type: string
      - description: Version (URL-encoded)
        in: path
        name: version
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "204":
          description: No content
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not a managed registry
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Server version not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Delete server version from specific registry
      tags:
      - registry
      - official
...
  /registry/{registryName}/v0.1/publish:
    post:
      description: Publish a server to the registry
      parameters:
      - description: Registry name
        in: path
        name: registryName
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "501":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not implemented
      security:
      - BearerAuth: []
      summary: Publish server
      tags:
      - registry
      - official
  /registry/{registryName}/v0.1/servers:
    get:
      description: Get a list of available servers in the registry
      parameters:
      - description: Registry name
        in: path
        name: registryName
        required: true
        schema:
          type: string
      - description: Pagination cursor for retrieving next set of results
        in: query
        name: cursor
        schema:
          type: string
      - description: Maximum number of items to return
        in: query
        name: limit
        schema:
          type: integer
      - description: Search servers by name (substring match)
        in: query
        name: search
        schema:
          type: string
      - description: Filter by version ('latest' for latest version, or an exact version
          like '1.2.3')
        in: query
        name: version
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v0.ServerListResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: List servers
      tags:
      - registry
      - official
  /registry/{registryName}/v0.1/servers/{serverName}/versions:
    get:
      description: Returns all available versions for a specific MCP server, ordered
        by publication date (newest first)
      parameters:
      - description: Registry name
        in: path
        name: registryName
        required: true
        schema:
          type: string
      - description: URL-encoded server name (e.g., \
        in: path
        name: serverName
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v0.ServerListResponse'
          description: A list of all versions for the server
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Server not found
      security:
      - BearerAuth: []
      summary: List all versions of an MCP server
      tags:
      - registry
      - official
  /registry/{registryName}/v0.1/servers/{serverName}/versions/{version}:
    get:
      description: |-
        Returns detailed information about a specific version of an MCP server.
        Use the special version `latest` to get the latest version.
      parameters:
      - description: Registry name
        in: path
        name: registryName
        required: true
        schema:
          type: string
      - description: URL-encoded server name (e.g., \
        in: path
        name: serverName
        required: true
        schema:
          type: string
      - description: URL-encoded version to retrieve (e.g., \
        in: path
        name: version
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v0.ServerResponse'
          description: Detailed server information
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Server or version not found
      security:
      - BearerAuth: []
      summary: Get specific MCP server version
      tags:
      - registry
      - official
  /registry/v0.1/publish:
    post:
      description: Publish a server to the registry
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "501":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not implemented
      security:
      - BearerAuth: []
      summary: Publish server
      tags:
      - registry
      - official
  /registry/v0.1/servers:
    get:
      description: Get a list of available servers in the registry
      parameters:
      - description: Pagination cursor for retrieving next set of results
        in: query
        name: cursor
        schema:
          type: string
      - description: Maximum number of items to return
        in: query
        name: limit
        schema:
          type: integer
      - description: Search servers by name (substring match)
        in: query
        name: search
        schema:
          type: string
      - description: Filter by version ('latest' for latest version, or an exact version
          like '1.2.3')
        in: query
        name: version
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v0.ServerListResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: List servers
      tags:
      - registry
      - official
  /registry/v0.1/servers/{serverName}/versions:
    get:
      description: Returns all available versions for a specific MCP server, ordered
        by publication date (newest first)
      parameters:
      - description: URL-encoded server name (e.g., \
        in: path
        name: serverName
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v0.ServerListResponse'
          description: A list of all versions for the server
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Server not found
      security:
      - BearerAuth: []
      summary: List all versions of an MCP server
      tags:
      - registry
      - official
  /registry/v0.1/servers/{serverName}/versions/{version}:
    get:
      description: |-
        Returns detailed information about a specific version of an MCP server.
        Use the special version `latest` to get the latest version.
      parameters:
      - description: URL-encoded server name (e.g., \
        in: path
        name: serverName
        required: true
        schema:
          type: string
      - description: URL-encoded version to retrieve (e.g., \
        in: path
        name: version
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v0.ServerResponse'
          description: Detailed server information
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Server or version not found
      security:
      - BearerAuth: []
      summary: Get specific MCP server version
      tags:
      - registry
      - official
...